+++ scripts/cljs-followers.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-# How to run multiple builds? Right now, I just start multiple processes
-clj -R:figwheel -m figwheel.main --build followers --repl
-